sistemata la logica di security e ruoli per vault

+++ src/main/java/osiride/vitt_be/service/VaultService.java
@@ -42,13 +42,24 @@ public class VaultService {
 
 	@Autowired
 	private AuthService authService;
-	
+
 	/**
 	 * Get all vaults in database
 	 * @return List<VaultDTO>
+	 * @throws BadRequestException 
+	 * @throws NotFoundException 
+	 * @throws InvalidTokenException 
+	 * @throws NotAuthorizedException 
 	 */
-	public List<VaultDTO> getAll(){
-		return vaultRepository.findAll().stream().map(vault -> vaultMapper.toDto(vault)).toList(); 
+	public List<VaultDTO> getAll() throws InvalidTokenException, NotFoundException, BadRequestException, NotAuthorizedException{
+		if(authService.isAdmin()) 
+		{
+			return vaultRepository.findAll().stream().map(vault -> vaultMapper.toDto(vault)).toList(); 
+		}
+		else {
+			log.error("SERVICE - Vault Operation not allowed - DELETE");
+			throw new NotAuthorizedException();
+		}
 	}
 
 	/**
@@ -68,12 +79,13 @@ public class VaultService {
 
 		if(maybeVault.isPresent()) {
 			VaultDTO vaultDTO = vaultMapper.toDto(maybeVault.get());
-			
+
 			UserDTO principal = authService.getPrincipal();
-			if(isOwner(principal, vaultDTO)) {
+			if(authService.isAdmin() || isOwner(principal, vaultDTO)) {
 				return vaultDTO;
 			}
 			else {
+				log.error("SERVICE - Vault Operation not allowed - FIND ONE");
 				throw new NotAuthorizedException();
 			}
 		}
@@ -90,18 +102,25 @@ public class VaultService {
 	 * @throws BadRequestException
 	 * @throws NotFoundException
 	 * @throws DuplicatedValueException 
+	 * @throws InvalidTokenException 
 	 * @throws SQLException
 	 */
-	
+
 	//FIXME CREARE UN METODO CHE CERCA SE IL VAULT ESISTE GIA CON QUEL NOME PER NON ROMPERE NEL DB
-	public VaultDTO create(VaultDTO vaultDTO) throws BadRequestException, NotFoundException, DuplicatedValueException{
+	public VaultDTO create(VaultDTO vaultDTO) throws BadRequestException, NotFoundException, DuplicatedValueException, InvalidTokenException{
 		if(vaultDTO == null || !isDataValid(vaultDTO)) {
 			log.error("SERVICE - Vault Data given is null - CREATE");
 			throw new BadRequestException();
 		}
 
 		Vault vault = vaultMapper.toEntity(vaultDTO);
-		vault.setUser(userMapper.toEntity(userService.findById(vaultDTO.getUserDTO().getId())));
+		
+		UserDTO user = authService.isAdmin() ? 
+                userService.findById(vaultDTO.getUserDTO().getId()) : 
+                authService.getPrincipal();
+		
+		vault.setUser(userMapper.toEntity(user));
+
 		vault.setId(null);
 		try {			
 			return vaultMapper.toDto(vaultRepository.save(vault));
@@ -118,12 +137,19 @@ public class VaultService {
 	 * @throws BadRequestException
 	 * @throws NotFoundException
 	 * @throws DuplicatedValueException
+	 * @throws InvalidTokenException 
+	 * @throws NotAuthorizedException 
 	 */
-	
-	public VaultDTO update(VaultDTO vaultDTO) throws BadRequestException, NotFoundException,DuplicatedValueException{
+
+	public VaultDTO update(VaultDTO vaultDTO) throws BadRequestException, NotFoundException ,DuplicatedValueException , InvalidTokenException, NotAuthorizedException{
 		if(vaultDTO == null || vaultDTO.getId() == null || !isDataValid(vaultDTO)) {
 			log.error("SERVICE - Vault Data given is null - UPDATE");
 			throw new BadRequestException();
+		}
+		
+		if(!(authService.isAdmin() || isOwner(authService.getPrincipal(),vaultDTO))) {
+			log.error("SERVICE - Vault Operation not allowed - UPDATE");
+			throw new NotAuthorizedException();
 		}
 
 		Optional<Vault> maybeVault = vaultRepository.findById(vaultDTO.getId());
@@ -133,16 +159,15 @@ public class VaultService {
 		}
 
 		Vault newVault = vaultMapper.toEntity(vaultDTO);
-
 		try {		
 			newVault= vaultRepository.save(newVault);
 		} catch (Exception e) {
+			e.printStackTrace();
 			log.error("SERVICE - Dupicated Vault Name - UPDATE");
 			throw new DuplicatedValueException();
 		}
 
 		return vaultMapper.toDto(newVault);
-
 	}
 
 	/**
@@ -152,9 +177,10 @@ public class VaultService {
 	 * @throws BadRequestException
 	 * @throws InternalServerException
 	 * @throws NotFoundException
+	 * @throws InvalidTokenException 
+	 * @throws NotAuthorizedException 
 	 */
-
-	public VaultDTO deleteById(Long id) throws BadRequestException, InternalServerException, NotFoundException{
+	public VaultDTO deleteById(Long id) throws BadRequestException, InternalServerException, NotFoundException, InvalidTokenException, NotAuthorizedException{
 		if(id == null) {
 			log.error("SERVICE - Vault id is null - DELETE");
 			throw new BadRequestException();
@@ -163,7 +189,14 @@ public class VaultService {
 		Optional<Vault> maybeVault = vaultRepository.findById(id);
 		if(maybeVault.isPresent()) {
 			Vault vault = maybeVault.get();
-			vaultRepository.delete(vault);
+
+			if(authService.isAdmin() || isOwner(authService.getPrincipal(), vaultMapper.toDto(vault))) {
+				vaultRepository.delete(vault);
+			}
+			else {
+				log.error("SERVICE - Vault Operation not allowed - DELETE");
+				throw new NotAuthorizedException();
+			}
 
 			if(!vaultRepository.existsById(id)) {
 				return vaultMapper.toDto(vault);
@@ -185,14 +218,22 @@ public class VaultService {
 	 * @return
 	 * @throws BadRequestException
 	 * @throws NotFoundException
+	 * @throws InvalidTokenException 
+	 * @throws NotAuthorizedException 
 	 */
-	public List<VaultDTO> getAllVaultByUserId(Long userId) throws BadRequestException, NotFoundException{
+	public List<VaultDTO> getAllVaultByUserId(Long userId) throws BadRequestException, NotFoundException, InvalidTokenException, NotAuthorizedException {
 		if(userId == null) {
-			log.error("SERVICE - Vault id is null - DELETE");
+			log.error("SERVICE - Vault id is null - GET VAUTLS BY USER");
 			throw new BadRequestException();
 		}
 
 		UserDTO user = userService.findById(userId);
+
+		if(!( authService.isAdmin() || authService.isPrincipal(userMapper.toEntity(user)))) {
+			log.error("SERVICE - Vault Operation not allowed - GET VAULTS BY USER");
+			throw new NotAuthorizedException();
+		}
+
 		List<VaultDTO> vaultList = 
 				vaultRepository.getAllByUser(userMapper.toEntity(user))
 				.stream()
@@ -200,7 +241,7 @@ public class VaultService {
 				.toList();
 		return vaultList;
 	}
-	
+
 	public List<VaultDTO> getAllVaultByPrincipal() throws BadRequestException, NotFoundException, InvalidTokenException{
 		UserDTO user = authService.getPrincipal();
 
@@ -286,12 +327,12 @@ public class VaultService {
 				? false 
 						: true;
 	}
-	
+
 	private boolean isOwner(UserDTO userDTO , VaultDTO vaultDTO) throws BadRequestException {
 		if(userDTO == null || vaultDTO == null) {
 			throw new BadRequestException();
 		}
-		
+
 		return vaultDTO.getUserDTO().equals(userDTO);
 	}
